+++ starkinfra/issuinginvoice/__issuinginvoice.py
@@ -65,7 +65,7 @@ def get(id, user=None):
 
 def query(limit=None, after=None, before=None, status=None, tags=None, user=None):
     """# Retrieve IssuingInvoices
-    Receive a generator of IssuingInvoices objects previously created in the Stark Infra API
+    Receive a generator of IssuingInvoice objects previously created in the Stark Infra API
     ## Parameters (optional):
     - limit [integer, default None]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
     - after [datetime.date or string, default None] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
@@ -74,7 +74,7 @@ def query(limit=None, after=None, before=None, status=None, tags=None, user=None
     - tags [list of strings, default None]: tags to filter retrieved objects. ex: ["tony", "stark"]
     - user [Organization/Project object, default None]: Organization or Project object. Not necessary if starkinfra.user was set before function call
     ## Return:
-    - generator of IssuingInvoices objects with updated attributes
+    - generator of IssuingInvoice objects with updated attributes
     """
     return rest.get_stream(
         resource=_resource,
@@ -89,7 +89,7 @@ def query(limit=None, after=None, before=None, status=None, tags=None, user=None
 
 def page(cursor=None, limit=None, after=None, before=None, status=None, tags=None, user=None):
     """# Retrieve IssuingInvoices
-    Receive a list of IssuingInvoices objects previously created in the Stark Infra API and the cursor to the next page.
+    Receive a list of IssuingInvoice objects previously created in the Stark Infra API and the cursor to the next page.
     ## Parameters (optional):
     - cursor [string, default None]: cursor returned on the previous page function call
     - limit [integer, default 100]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
@@ -99,8 +99,8 @@ def page(cursor=None, limit=None, after=None, before=None, status=None, tags=Non
     - tags [list of strings, default None]: tags to filter retrieved objects. ex: ["tony", "stark"]
     - user [Organization/Project object, default None]: Organization or Project object. Not necessary if starkinfra.user was set before function call
     ## Return:
-    - list of IssuingInvoices objects with updated attributes
-    - cursor to retrieve the next page of IssuingInvoices objects
+    - list of IssuingInvoice objects with updated attributes
+    - cursor to retrieve the next page of IssuingInvoice objects
     """
     return rest.get_page(
         resource=_resource,