+++ starkinfra/issuingpurchase/__issuingpurchase.py
@@ -1,8 +1,9 @@
 from json import dumps
-from starkcore.utils.resource import Resource
-from starkcore.utils.checks import check_datetime, check_date
 from ..utils import rest
 from ..utils.parse import parse_and_verify
+from starkcore.utils.api import api_json
+from starkcore.utils.resource import Resource
+from starkcore.utils.checks import check_datetime, check_date
 
 
 class IssuingPurchase(Resource):
@@ -98,7 +99,7 @@ def get(id, user=None):
 def query(ids=None, limit=None, after=None, before=None, end_to_end_ids=None, holder_ids=None, card_ids=None,
           status=None, user=None):
     """# Retrieve IssuingPurchase
-    Receive a generator of IssuingPurchases objects previously created in the Stark Infra API
+    Receive a generator of IssuingPurchase objects previously created in the Stark Infra API
     ## Parameters (optional):
     - ids [list of strings, default [], default None]: purchase IDs
     - limit [integer, default None]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
@@ -127,7 +128,7 @@ def query(ids=None, limit=None, after=None, before=None, end_to_end_ids=None, ho
 
 
 def page(end_to_end_ids=None, holder_ids=None, card_ids=None, status=None, after=None, before=None, ids=None,
-         cursor=None, limit=None, user=None):
+        cursor=None, limit=None, user=None):
     """# Retrieve paged IssuingPurchase
     Receive a list of IssuingPurchase objects previously created in the Stark Infra API and the cursor to the next page.
     ## Parameters (optional):
@@ -191,14 +192,15 @@ def response(status, amount=None, reason=None, tags=None):
     ## Parameters (conditionally required):
     - reason [string]: denial reason. Options: "other", "blocked", "lostCard", "stolenCard", "invalidPin", "invalidCard", "cardExpired", "issuerError", "concurrency", "standInDenial", "subIssuerError", "invalidPurpose", "invalidZipCode", "invalidWalletId", "inconsistentCard", "settlementFailed", "cardRuleMismatch", "invalidExpiration", "prepaidInstallment", "holderRuleMismatch", "insufficientBalance", "tooManyTransactions", "invalidSecurityCode", "invalidPaymentMethod", "confirmationDeadline", "withdrawalAmountLimit", "insufficientCardLimit", "insufficientHolderLimit"
     ## Parameters (optional):
-    - amount [integer, default 0]: amount in cents that was authorized. ex: 1234 (= R$ 12.34)
-    - tags [list of strings, default []]: tags to filter retrieved object. ex: ["tony", "stark"]
+    - amount [integer, default None]: amount in cents that was authorized. ex: 1234 (= R$ 12.34)
+    - tags [list of strings, default None]: tags to filter retrieved object. ex: ["tony", "stark"]
     ## Return:
     - Dumped JSON string that must be returned to us on the IssuingPurchase request
     """
-    return dumps({"authorization": {
+    params = {"authorization": {
         "status": status,
-        "amount": amount or 0,
-        "reason": reason or "",
-        "tags": tags or [],
-    }})
+        "amount": amount,
+        "reason": reason,
+        "tags": tags,
+    }}
+    return dumps(api_json(params))