+++ starkinfra/pixinfraction/__pixinfraction.py
@@ -80,7 +80,7 @@ def get(id, user=None):
 
 def query(limit=None, after=None, before=None, status=None, ids=None, type=None, user=None):
     """# Retrieve PixInfractions
-    Receive a generator of PixInfractions objects previously created in the Stark Infra API
+    Receive a generator of PixInfraction objects previously created in the Stark Infra API
     ## Parameters (optional):
     - limit [integer, default None]: maximum number of objects to be retrieved. Unlimited if None. ex: 35
     - after [datetime.date or string, default None]: date filter for objects created after a specified date. ex: datetime.date(2020, 3, 10)
@@ -108,7 +108,7 @@ def query(limit=None, after=None, before=None, status=None, ids=None, type=None,
 def page(cursor=None, limit=None, after=None, before=None, status=None, ids=None, type=None,
          user=None):
     """# Retrieve paged PixInfractions
-    Receive a list of up to 100 PixInfractions objects previously created in the Stark Infra API and the cursor to the next page.
+    Receive a list of up to 100 PixInfraction objects previously created in the Stark Infra API and the cursor to the next page.
     Use this function instead of query if you want to manually page your requests.
     ## Parameters (optional):
     - cursor [string, default None]: cursor returned on the previous page function call.